# pytorrent/openapi/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "error": {
            "type": "string"
          },
          "ok": {
            "enum": [
              false
            ],
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "error"
        ],
        "type": "object"
      },
      "ApiOk": {
        "properties": {
          "ok": {
            "enum": [
              true
            ],
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "type": "object"
      },
      "ApiToken": {
        "properties": {
          "created_at": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "last_used_at": {
            "nullable": true,
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "revoked_at": {
            "nullable": true,
            "type": "string"
          },
          "token_prefix": {
            "type": "string"
          },
          "user_id": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "user_id",
          "name",
          "token_prefix",
          "created_at"
        ],
        "type": "object"
      },
...
      "AuthUser": {
        "additionalProperties": true,
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "id": {
            "type": "integer"
          },
          "role": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "apiKeyAuth": {
        "description": "Optional per-user API token. Tokens are accepted only when PYTORRENT_AUTH_ENABLE=1. When auth is disabled, API endpoints do not require credentials.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "bearerApiToken": {
        "description": "Optional per-user API token in Authorization: Bearer <token>. Tokens are accepted only when PYTORRENT_AUTH_ENABLE=1.",
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "description": "Optional Flask session cookie. Required only when PYTORRENT_AUTH_ENABLE=1 and no API token is used.",
        "in": "cookie",
        "name": "session",
        "type": "apiKey"
      }
    },
    "parameters": {
      "ProfileId": {
        "name": "profile_id",
        "in": "query",
        "required": false,
        "description": "Optional rTorrent profile id for profile-scoped API operations. If omitted, the user active profile is used for backward compatibility; fallback profile id 1 is used only when no active profile exists and the user can access it.",
        "schema": {
          "type": "integer",
          "default": 1,
          "minimum": 1
        }
      },
      "ProfileName": {
        "name": "profile_name",
        "in": "query",
        "required": false,
        "description": "Optional rTorrent profile name for profile-scoped API operations. profile_id has priority when both are supplied.",
        "schema": {
          "type": "string"
        }
      }
    }
  },
  "info": {
    "title": "pyTorrent API",
    "version": "0.0.1"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "password": {
                    "format": "password",
                    "type": "string"
                  },
                  "username": {
                    "type": "string"
                  },
                  "profile_id": {
                    "type": "integer",
                    "default": 1,
                    "minimum": 1,
                    "description": "Optional rTorrent profile id. Overrides active profile for this request."
                  },
                  "profile_name": {
                    "type": "string",
                    "description": "Optional rTorrent profile name. Used only when profile_id is not provided."
                  }
                },
                "required": [
                  "username",
                  "password"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "auth_enabled": {
                          "type": "boolean"
                        },
                        "user": {
                          "$ref": "#/components/schemas/AuthUser"
                        }
                      },
                      "required": [
                        "user",
                        "auth_enabled"
                      ],
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid credentials"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication disabled"
          }
        },
        "summary": "Log in with username and password",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiOk"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication disabled"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Log out current user",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "auth_enabled": {
                          "type": "boolean"
                        },
                        "user": {
                          "$ref": "#/components/schemas/AuthUser"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication disabled"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Get current user",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
    "/api/auth/users": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "users": {
                          "items": {
                            "$ref": "#/components/schemas/AuthUser"
                          },
                          "type": "array"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication disabled"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "List users",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "properties": {
                  "profile_id": {
                    "type": "integer",
                    "default": 1,
                    "minimum": 1,
                    "description": "Optional rTorrent profile id. Overrides active profile for this request."
                  },
                  "profile_name": {
                    "type": "string",
                    "description": "Optional rTorrent profile name. Used only when profile_id is not provided."
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/AuthUser"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Create user",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ]
      }
    },
    "/api/auth/users/{user_id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "users": {
                          "items": {
                            "$ref": "#/components/schemas/AuthUser"
                          },
                          "type": "array"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Delete user"
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "properties": {
                  "profile_id": {
                    "type": "integer",
                    "default": 1,
                    "minimum": 1,
                    "description": "Optional rTorrent profile id. Overrides active profile for this request."
                  },
                  "profile_name": {
                    "type": "string",
                    "description": "Optional rTorrent profile name. Used only when profile_id is not provided."
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/AuthUser"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Update user"
      }
    },
    "/api/auth/users/{user_id}/tokens": {
      "get": {
        "description": "Available only when application authentication is enabled. Admins can list any user tokens; regular users can list their own tokens. Token secrets are never returned.",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "tokens": {
                      "items": {
                        "$ref": "#/components/schemas/ApiToken"
                      },
                      "type": "array"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Token list"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "apiKeyAuth": []
          },
          {
            "bearerApiToken": []
          }
        ],
        "summary": "List API tokens for a user"
      },
      "post": {
        "description": "Available only when application authentication is enabled. The returned token is shown once and must be copied immediately.",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "name": {
                    "example": "automation",
                    "type": "string"
                  },
                  "profile_id": {
                    "type": "integer",
                    "default": 1,
                    "minimum": 1,
                    "description": "Optional rTorrent profile id. Overrides active profile for this request."
                  },
                  "profile_name": {
                    "type": "string",
                    "description": "Optional rTorrent profile name. Used only when profile_id is not provided."
                  }
                },
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "token": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/ApiToken"
                        },
                        {
                          "properties": {
                            "token": {
                              "type": "string"
                            }
                          },
                          "type": "object"
                        }
                      ]
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Generated token"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "apiKeyAuth": []
          },
          {
            "bearerApiToken": []
          }
        ],
        "summary": "Generate API token for a user"
      }
    },
    "/api/auth/users/{user_id}/tokens/{token_id}": {
      "delete": {
        "description": "Available only when application authentication is enabled. Revoked tokens stop working immediately.",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "token_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "$ref": "#/components/parameters/ProfileId"
          },
          {
            "$ref": "#/components/parameters/ProfileName"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "tokens": {
                      "items": {
                        "$ref": "#/components/schemas/ApiToken"
                      },
                      "type": "array"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Updated token list"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "apiKeyAuth": []
          },
          {
            "bearerApiToken": []
          }
        ],
        "summary": "Revoke API token"
      }
    },
...
  }
}